VM dry-run: define a specification structure exercised with exit and move
For every instruction:
  * Define a specific test representing its specification (in vm/is)
  * Provide a collection of helpers for the fetch stage

Helpers allow to fetch source and destination operands according to their
expected profile. Right now, these are static definitions per profile,
with the intention to be enriched on demand. We may envision to have a
generic class to do this, with clear profile definitions, but it's not
necessarily a plus.