VM: start prototyping Compare instruction
Purpose of Compare is to store the result of a comparison of two
operands into a destination operand, for example: z = (x == y).
This instruction is not super-useful in itself, but is the root of
jump conditional and select operations.

This introduces the Comparator ALU block, that performs standard
arithmetic comparisons: ==, !=, <, >, <= and >=.

Comparators are essentially unimplemented in this commit. The main
objective is to implement generic functions strictly complying with
the C standard, exposed in section 6.3.1.8, stored in the misc directory.