VM: Start prototyping an ALU
!!!The code is not tested and very probably wrong!!!

Principle is to map values to "VM registers", i.e. an internal
representation able to map signed and unsigned values with different
dynamics.
Right now, there is only an instance of VmRegister for any length of
integer (LargeNumber), but the final goal is to have more efficient
implementations, depending on the integer size (e.g. 64 bits for 32 bits
integers...).
The ALU first "nomalizes" (normalize and equalize functions) operands
to be sure to operate with the same types of VM registers, then calls
the associated operation (e.g. add, compare...) using
VmRegisterOperations objects for the given type.