VM: fixing bug while starting to test bitflip
Found an issue when specifying -1:u32.
The problem is that BigInt is not reliable for the kind of operation we would like to achieve.
A BigInt is, for excellent reason, compacted to the minimum number of bytes.
So, toByteArray of a translation of -1:u32 translated to BigInt, returns a single byte equal to -1: dynamic is lost.
Force to use as much as possible byte arrays, with few converters to help (some shall be removed in a future version).