VM/ASM: start testing comparisons
Test disabled upon push, since I need to fix the bugs.
First set of tests consists of verifying that the VM is consistent
with the C standard definition, considering two numbers with the
same encoding length.
Most critical part is to properly handle signed/unsigned consistency
when the two operands have different signedness.
Right now it fails.

Notice that the test bench is generated by a bit of C code (cmptest16.c),
so we consider that gcc gives the truth in any case.