VM: only support "standard" machine types
Exotic types with byte length different from a power of two are allowed
by stdint, but "non-standard". Looking at different implementations,
they clearly appear to have machine specific implementations, which
cannot be supported in a universal framework.

As a consequence, if there is a need to support this, then it must
be handled at a much higher level than the VM.